backend/requirements.txt

fastapi
uvicorn
gunicorn
python-multipart
pillow
imagehash
